+++ cmd/fuji/main.go
@@ -74,7 +74,7 @@ func parseRunCommandArgs(args []string) (src string, noOpt bool, err error) {
 }
 
 // version is set by release builds, e.g. -ldflags "-X main.version=1.0.0"
-var version = "0.3.0"
+var version = "0.3.1"
 
 func main() {
 	args := os.Args[1:]